Naive implementation of a cluster using netty
With this change we should be able to run multiple nodes with something
like passed by argument

```
8081 127.0.0.1:8080 127.0.0.1:8082
```

This will inform the application that it should use the port 8081 and
the other members of the cluster are 127.0.0.1:8080 and 127.0.0.1:8082.

It handles reconnection, so if a node crashes and it will try to
reconnect until the node is up again.