Remove the need for the EchoServer
The communication between the server and the client should be
transparent to the user of the cluster. All the information needs to be
in a layer specific for networking.

This is the first step to isolate that layer and transform a
client-server into a messaging architecture. The end goal is to have
messages communicating between the members of the cluster.