Remove the method connectToNode
The connection will be managed while trying to send a message to a
specific address, therefore we don't need to connect to a specific host
or deal with connection closed after the initial connect.

It needs some refactoring, but I think it should be done while doing the
pool of connections.